refactor : 접속자 수 증감 구현 위치 변경 #74
