+++ src/main/java/com/sparta/spartatigers/domain/liveboard/interceptor/LiveBoardInterceptor.java
@@ -25,7 +25,7 @@ public class LiveBoardInterceptor implements ChannelInterceptor {
                 StompHeaderAccessor.wrap(message); // stomp 메세지의 헤더를 분석 ( 커멘드, 세션아이디 등등..)
         StompCommand command = accessor.getCommand();
 
-        // 채팅방 구독시
+		// 채팅방 구독시
         if (StompCommand.SUBSCRIBE.equals(command)) {
             String destination = accessor.getDestination(); // url
             String sessionId = accessor.getSessionId(); // 세션 id
@@ -40,7 +40,6 @@ public class LiveBoardInterceptor implements ChannelInterceptor {
                 }
 
                 // 새로 입장한 채팅방 접속자 수 증가
-                liveBoardRoomService.increaseConnectCount(enterRoomId);
                 redisTemplate.opsForValue().set(sessionId, enterRoomId);
             }
         }
@@ -51,7 +50,6 @@ public class LiveBoardInterceptor implements ChannelInterceptor {
             String roomId = redisTemplate.opsForValue().get(sessionId);
 
             if (roomId != null) {
-                liveBoardRoomService.decreaseConnectCount(roomId);
                 redisTemplate.delete(sessionId);
             }
         }
