+++ src/main/java/com/sparta/spartatigers/domain/liveboard/model/LiveBoardRoom.java
@@ -17,13 +17,7 @@ public class LiveBoardRoom {
     private Long matchId;
     private String title;
     private LocalDateTime openAt;
-    private LocalDateTime closedAt;
-    // private boolean isClosed;
     private int connectCount;
-
-    // public void close() {
-    //     this.isClosed = true;
-    // }
 
     public void increaseCount() {
         this.connectCount++;
